# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Acme SaaS API",
    "version": "1.0.0",
    "description": "RESTful API for plan selection and account onboarding. Each response includes a `nextAction` field that indicates the next endpoint to call, allowing callers to complete the signup funnel sequentially without embedding flow logic on the client.\n\n**Prefer GET over POST.** Discovery endpoints (`/api/pricing`, `/api/plans`, `/api/recommend-plan`) all support GET with query parameters and work from any HTTP client including browser-based LLM tools. Only use POST for `recommend-plan` when query parameters are not available. Account creation, subscription, and checkout are POST-only because they create server-side state.\n\n**Static fallback for read-only agents:** If you cannot construct dynamic URLs or make POST requests, fetch `https://llmsopenapi.vercel.app/api/pricing` (no parameters needed) to get all plans with prices and seat ranges. Match the user's requirements (team size, SSO need) against the returned `minEmployees`/`maxEmployees` fields to recommend a plan. Direct the user to complete signup at `https://llmsopenapi.vercel.app/signup` or view plans at `https://llmsopenapi.vercel.app/features`.",
    "contact": {
      "name": "Acme SaaS",
      "url": "https://llmsopenapi.vercel.app"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
...
  ],
...
  "components": {
    "schemas": {
      "PlanId": {
        "type": "string",
        "enum": ["starter", "advantage", "premier", "enterprise"],
        "description": "Machine-readable plan identifier. `starter` is the free Starter plan (API-only, not shown on the /features page). Use the value returned by `/api/recommend-plan` when calling `/api/create-subscription`.",
        "example": "enterprise"
      },
...
    },
...
  },
  "paths": {
    "/api/pricing": {
      "get": {
        "operationId": "getPricing",
        "summary": "Get pricing for all products",
        "description": "Returns the product catalogue with display names matching the /features page. Each product has a `prices` array with one entry per supported country. Pass `country` to flag the relevant price as `is_default: true`. Use `planId` from each product when calling `/api/create-subscription`.",
        "tags": ["Onboarding"],
        "parameters": [
          {
            "name": "country",
            "in": "query",
            "required": false,
            "schema": { "type": "string" },
            "description": "ISO country code (e.g. `CR`). Marks the matching price entry `is_default: true`.",
            "example": "CR"
          },
          {
            "name": "employees",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "minimum": 1 },
            "description": "Number of employees. When provided alongside `needsSSO`, the matching product is flagged `is_recommended: true`.",
            "example": 4
          },
          {
            "name": "needsSSO",
            "in": "query",
            "required": false,
            "schema": { "type": "boolean" },
            "description": "Whether SSO is required. Used together with `employees` to flag the best-fit product.",
            "example": true
          }
        ],
        "responses": {
          "200": {
            "description": "Product catalogue",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "products": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "sku":    { "type": "string", "example": "team-advantage" },
                          "name":   { "type": "string", "example": "Example Standard" },
                          "active": { "type": "boolean" },
                          "planId":           { "$ref": "#/components/schemas/PlanId" },
                          "minEmployees":     { "type": "integer", "description": "Minimum team size for this plan.", "example": 4 },
                          "maxEmployees":     { "oneOf": [{ "type": "integer" }, { "type": "string", "enum": ["unlimited"] }], "description": "Maximum team size. `\"unlimited\"` for Team Enterprise.", "example": 10 },
                          "responsesPerYear": { "type": "integer", "description": "Annual response allowance.", "example": 50000 },
                          "is_recommended":   { "type": "boolean", "description": "Present only when `employees` is passed. `true` on the product that best fits the given criteria." },
                          "prices": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "country":        { "type": "string", "example": "US" },
                                "currency":       { "type": "string", "example": "usd" },
                                "unit_amount":    { "type": "integer", "description": "Price in minor currency units (cents)", "example": 1000 },
                                "billing_scheme": { "type": "string", "example": "per_unit" },
                                "type":           { "type": "string", "example": "recurring" },
                                "is_default":     { "type": "boolean" }
                              }
                            }
                          }
                        }
                      }
                    },
                    "plansUrl":  { "type": "string", "format": "uri" },
                    "signupUrl": { "type": "string", "format": "uri" }
                  }
                },
                "example": {
                  "products": [
                    {
                      "sku": "starter", "name": "Example Starter", "active": true, "planId": "starter", "minEmployees": 1, "maxEmployees": 3, "responsesPerYear": 5000,
                      "prices": [
                        { "country": "US", "currency": "usd", "unit_amount": 0, "billing_scheme": "per_unit", "type": "recurring", "is_default": true },
                        { "country": "CR", "currency": "crc", "unit_amount": 0, "billing_scheme": "per_unit", "type": "recurring", "is_default": false }
                      ]
                    },
                    {
                      "sku": "team-advantage", "name": "Example Standard", "active": true, "planId": "advantage", "minEmployees": 4, "maxEmployees": 10, "responsesPerYear": 50000,
                      "prices": [
                        { "country": "US", "currency": "usd", "unit_amount": 1000, "billing_scheme": "per_unit", "type": "recurring", "is_default": true },
                        { "country": "CR", "currency": "crc", "unit_amount": 5000, "billing_scheme": "per_unit", "type": "recurring", "is_default": false }
                      ]
                    },
                    {
                      "sku": "team-premier", "name": "Example Pro", "active": true, "planId": "premier", "minEmployees": 11, "maxEmployees": 50, "responsesPerYear": 100000,
                      "prices": [
                        { "country": "US", "currency": "usd", "unit_amount": 2000, "billing_scheme": "per_unit", "type": "recurring", "is_default": true },
                        { "country": "CR", "currency": "crc", "unit_amount": 10000, "billing_scheme": "per_unit", "type": "recurring", "is_default": false }
                      ]
                    },
                    {
                      "sku": "team-enterprise", "name": "Example Max", "active": true, "planId": "enterprise", "minEmployees": 51, "maxEmployees": "unlimited", "responsesPerYear": 200000,
                      "prices": [
                        { "country": "US", "currency": "usd", "unit_amount": 5000, "billing_scheme": "per_unit", "type": "recurring", "is_default": true },
                        { "country": "CR", "currency": "crc", "unit_amount": 25000, "billing_scheme": "per_unit", "type": "recurring", "is_default": false }
                      ]
                    }
                  ],
                  "plansUrl":  "https://llmsopenapi.vercel.app/features",
                  "signupUrl": "https://llmsopenapi.vercel.app/signup"
                }
              }
            }
          }
        }
      }
    },
    "/api/plans": {
      "get": {
        "operationId": "listPlans",
        "summary": "List all plans",
        "description": "Returns all available plans with pricing. Use this for discovery — e.g. showing a user their options before asking for company details. Pass `country` as a query parameter to get localised pricing. Includes `signupUrl` and `plansUrl` in the response.",
        "tags": ["Onboarding"],
        "parameters": [
          {
            "name": "country",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "minLength": 2, "maxLength": 3, "example": "CR" },
            "description": "ISO 3166-1 alpha-2 or alpha-3 country code. Determines response currency. Defaults to USD if omitted."
          }
        ],
        "responses": {
          "200": {
            "description": "All available plans with pricing.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["plans", "signupUrl", "plansUrl"],
                  "properties": {
                    "plans": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Plan" }
                    },
                    "signupUrl": { "type": "string", "format": "uri", "example": "https://llmsopenapi.vercel.app/signup" },
                    "plansUrl": { "type": "string", "format": "uri", "example": "https://llmsopenapi.vercel.app/features" }
                  }
                },
                "example": {
                  "plans": [
                    { "planId": "starter",    "name": "Example Starter",  "description": "Free plan for individuals and very small teams.", "minEmployees": 1,  "maxEmployees": 3,           "supportsSSO": false, "price": 0,    "currency": "USD", "billingCycle": "monthly" },
                    { "planId": "advantage",  "name": "Example Standard", "description": "For small teams getting started.",               "minEmployees": 4,  "maxEmployees": 10,          "supportsSSO": false, "price": 1000, "currency": "USD", "billingCycle": "monthly" },
                    { "planId": "premier",    "name": "Example Pro",      "description": "For growing teams that need more capacity.",     "minEmployees": 11, "maxEmployees": 50,          "supportsSSO": false, "price": 2000, "currency": "USD", "billingCycle": "monthly" },
                    { "planId": "enterprise", "name": "Example Max",      "description": "For large teams with advanced needs including SSO.", "minEmployees": 51, "maxEmployees": "unlimited", "supportsSSO": true,  "price": 5000, "currency": "USD", "billingCycle": "monthly" }
                  ],
                  "signupUrl": "https://llmsopenapi.vercel.app/signup",
                  "plansUrl": "https://llmsopenapi.vercel.app/features"
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/api/recommend-plan": {
      "get": {
        "operationId": "recommendPlanGet",
        "summary": "Recommend a plan (preferred)",
        "description": "**Preferred method.** Accepts query parameters — works from any HTTP client including browser-based LLM tools (Claude.ai, ChatGPT, Gemini). Returns the best-fit plan with localized pricing and a `signupUrl` to continue. Use this unless your client cannot issue GET requests with query parameters.",
        "tags": ["Onboarding"],
        "parameters": [
          {
            "name": "employees",
            "in": "query",
            "required": true,
            "schema": { "type": "integer", "minimum": 1, "example": 15 },
            "description": "Number of employees who will use the platform."
          },
          {
            "name": "country",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "minLength": 2, "maxLength": 3, "example": "CR" },
            "description": "ISO 3166-1 country code. Determines response currency."
          },
          {
            "name": "needsSSO",
            "in": "query",
            "required": false,
            "schema": { "type": "boolean", "default": false, "example": true },
            "description": "Pass `true` if SSO is required. Forces Enterprise plan."
          }
        ],
        "responses": {
          "200": {
            "description": "Plan recommendation with localized pricing and signup link.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PlanRecommendation" },
                "example": {
                  "recommendedPlan": "Example Pro",
                  "planId": "premier",
                  "price": 2000,
                  "currency": "USD",
                  "billingCycle": "monthly",
                  "reasons": ["15 employees fits Example Pro (11–50 employees)", "SSO required — only Example Max supports SSO"],
                  "signupUrl": "https://llmsopenapi.vercel.app/signup",
                  "plansUrl": "https://llmsopenapi.vercel.app/features",
                  "nextAction": "create_account"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      },
      "post": {
        "operationId": "recommendPlan",
        "summary": "Recommend a plan (POST alternative)",
        "description": "**Use `GET /api/recommend-plan` instead when possible.** This POST variant exists for clients that cannot send query parameters. Accepts the same inputs as a JSON body and returns an identical response. The `planId` and pricing in the response are the inputs for subsequent steps; do not hard-code plan names or prices. Returns `nextAction: \"create_account\"`.",
        "tags": ["Onboarding"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["employees"],
                "properties": {
                  "country": {
                    "type": "string",
                    "description": "ISO 3166-1 alpha-2 country code. Determines the response currency. Defaults to USD if omitted or unrecognised.",
                    "minLength": 2,
                    "maxLength": 2,
                    "example": "CR"
                  },
                  "employees": {
                    "type": "integer",
                    "minimum": 1,
                    "description": "Number of employees who will use the platform. Determines the minimum eligible plan tier.",
                    "example": 15
                  },
                  "needsSSO": {
                    "type": "boolean",
                    "default": false,
                    "description": "Whether Single Sign-On (SSO) is required. When `true`, the Enterprise plan is always recommended regardless of employee count.",
                    "example": true
                  }
                }
              },
              "example": { "country": "CR", "employees": 15, "needsSSO": true }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Plan recommendation with localized pricing.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PlanRecommendation" },
                "example": {
                  "recommendedPlan": "Example Pro",
                  "planId": "premier",
                  "price": 2000,
                  "currency": "USD",
                  "billingCycle": "monthly",
                  "reasons": [
                    "15 employees fits Example Pro (11–50 employees)",
                    "SSO required — only Example Max supports SSO"
                  ],
                  "signupUrl": "https://llmsopenapi.vercel.app/signup",
                  "plansUrl": "https://llmsopenapi.vercel.app/features",
                  "nextAction": "create_account"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "405": { "$ref": "#/components/responses/MethodNotAllowed" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/api/create-account": {
      "post": {
        "operationId": "createAccount",
        "summary": "Create an account",
        "description": "Step 2. Register a new user account. Returns an `accountId` and `nextAction: \"create_subscription\"`. Returns HTTP 409 with the existing `accountId` if the email is already registered.",
        "tags": ["Onboarding"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "name", "company"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Contact email address. Must be unique — used as the account identifier.",
                    "example": "jane@acme.com"
                  },
                  "name": {
                    "type": "string",
                    "description": "Full name of the primary contact.",
                    "minLength": 1,
                    "example": "Jane Smith"
                  },
                  "company": {
                    "type": "string",
                    "description": "Company or organisation name.",
                    "minLength": 1,
                    "example": "Acme Corp"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created successfully.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Account" },
                "example": {
                  "accountId": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                  "email": "jane@acme.com",
                  "company": "Acme Corp",
                  "createdAt": "2026-07-05T14:00:00.000Z",
                  "nextAction": "create_subscription"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "405": { "$ref": "#/components/responses/MethodNotAllowed" },
          "409": {
            "description": "An account with this email already exists. The existing `accountId` is returned so the caller can continue the flow without creating a duplicate.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/Error" },
                    {
                      "type": "object",
                      "required": ["accountId"],
                      "properties": {
                        "accountId": {
                          "type": "string",
                          "format": "uuid",
                          "description": "ID of the existing account. Use this to proceed to `/api/create-subscription`.",
                          "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/api/create-subscription": {
      "post": {
        "operationId": "createSubscription",
        "summary": "Create a subscription",
        "description": "Step 3. Link an account to a plan. Returns a `subscriptionId`, full order summary, and `nextAction: \"confirm_checkout\"`. **The caller must present the order summary to the user and obtain explicit confirmation before calling `/api/create-checkout`.**",
        "tags": ["Onboarding"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["accountId", "planId"],
                "properties": {
                  "accountId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "Account ID returned by `/api/create-account`.",
                    "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                  },
                  "planId": {
                    "$ref": "#/components/schemas/PlanId"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Subscription created in `pending` state. Present the order summary to the user before proceeding.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Subscription" },
                "example": {
                  "subscriptionId": "7c9e6679-7425-40de-944b-e07fc1f90ae7",
                  "accountId": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                  "plan": "Example Pro",
                  "price": 2000,
                  "currency": "USD",
                  "billingCycle": "monthly",
                  "status": "pending",
                  "nextAction": "confirm_checkout",
                  "agentNote": "Present the order summary to the user and wait for explicit confirmation before calling /api/create-checkout."
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "405": { "$ref": "#/components/responses/MethodNotAllowed" },
          "409": {
            "description": "A pending subscription already exists for this account. The existing `subscriptionId` is returned so the caller can proceed directly to confirmation.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/Error" },
                    {
                      "type": "object",
                      "required": ["subscriptionId", "nextAction"],
                      "properties": {
                        "subscriptionId": {
                          "type": "string",
                          "format": "uuid",
                          "description": "ID of the existing pending subscription.",
                          "example": "7c9e6679-7425-40de-944b-e07fc1f90ae7"
                        },
                        "nextAction": {
                          "$ref": "#/components/schemas/NextAction"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
...
    "/api/reset": {
      "delete": {
        "operationId": "resetData",
        "summary": "Reset all data",
        "description": "Call before each test run or demo session to start from a clean state. Wipes all accounts, subscriptions, and checkouts from Redis. Do not call during an active onboarding flow — all in-progress data will be lost. Intended for development and demo use only.",
        "tags": ["Utility"],
        "responses": {
          "200": {
            "description": "All data cleared.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["message"],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "All data cleared."
                    }
                  }
                }
              }
            }
          },
          "405": { "$ref": "#/components/responses/MethodNotAllowed" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    }
  }
}